Ticket 31 : delete address, delete patch method, refactor role in abstractAccount for company and admin
