+++ src/Controller/ProductController.php
@@ -151,77 +151,6 @@ class ProductController extends AbstractController
      */
     #[Route(
         '/api/products/{id}',
-        name: 'patch_product',
-        methods: ['PATCH']
-    )]
-    public function updateProduct(
-        int $id,
-        Request $request,
-        ProductRepository $productRepository,
-        EntityManagerInterface $em,
-    ): JsonResponse {
-        $product = $productRepository->find($id);
-        if (!$product) {
-            return new JsonResponse(['message' => 'Produit non trouvé.'], Response::HTTP_NOT_FOUND);
-        }
-
-        $data = json_decode($request->getContent(), true);
-        if (JSON_ERROR_NONE !== json_last_error()) {
-            return new JsonResponse(['message' => 'Données JSON invalides.'], Response::HTTP_BAD_REQUEST);
-        }
-
-        if (isset($data['name'])) {
-            $product->setName($data['name']);
-        }
-        if (isset($data['description'])) {
-            $product->setDescription($data['description']);
-        }
-        if (isset($data['model'])) {
-            $product->setModel($data['model']);
-        }
-        if (isset($data['brand'])) {
-            $product->setBrand($data['brand']);
-        }
-        if (isset($data['reference'])) {
-            $product->setReference($data['reference']);
-        }
-        if (isset($data['price'])) {
-            $product->setPrice((float) $data['price']);
-        }
-        if (isset($data['dimension'])) {
-            $product->setDimension($data['dimension']);
-        }
-        if (isset($data['stock'])) {
-            $product->setStock((int) $data['stock']);
-        }
-        if (isset($data['isAvailable'])) {
-            $product->setIsAvailable((bool) $data['isAvailable']);
-        }
-        if (isset($data['image'])) {
-            $product->setImage($data['image']);
-        }
-
-        $errors = $this->getValidationErrors($product);
-        if (!empty($errors)) {
-            return new JsonResponse(['errors' => $errors], Response::HTTP_BAD_REQUEST);
-        }
-
-        $em->flush();
-
-        $this->cache->invalidateTags(['productCache']);
-        $this->cache->delete('getProductDetail-'.$id);
-
-        $serializationContext = SerializationContext::create()->setGroups(['product:read']);
-        $jsonContent = $this->serializer->serialize($product, 'json', $serializationContext);
-
-        return new JsonResponse($jsonContent, Response::HTTP_OK, [], true);
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    #[Route(
-        '/api/products/{id}',
         name: 'delete_product',
         methods: ['DELETE']
     )]
